Moduk CRUD Books (bagian CREATE) - Tampilan dan bagian backend selesai (tinggal bagian validasi javascript)

+++ database/migrations/2021_01_01_112311_create_books_table.php
@@ -27,7 +27,7 @@ class CreateBooksTable extends Migration
             $table->date('release_date');
             $table->string('publisher');
             $table->string('subtitle');
-            $table->float('weight');
+            $table->integer('weight');
             $table->float('width');
             $table->float('height');
             $table->timestamps();
